Enable null state analysis
Adjust code that causes the analysis to emit warnings to check
variables for null values before using them.

In this version, checks for values provided by the config system are
checked just before they are used. In the cases where a fallback value
can't be used, the program crashes. To avoid crashes in the middle of
running commands that e.g. writes files, we might want to prevalidate
the config before running the commands in the future.