Remove UserFileConfig
It was decided that at least for now, everything that's needed can be
achieved with env config, so the user level config stub was removed. It
might be implemented if the need arises in the future.

Since AuthConfig now only contains one token, it was renamed from
DefaultToken to AuthToken.