Show notification if newer version has been released
The check for update is initiated asynchronously before running the
actual command. If the check has completed when the command is done,
a notification about possible update is shown. If the check is
incomplete, the program exits instead of waiting the check to complete.

Release information is read from GitHub repo. The version number is
read from the JSON response with regex to avoid JSON deserialization.
This requires that the release versions follow the X.Y.Z semantic
versioning.

In case of errors while reading the version of the latest release, the
method fails silently, since it's unlikely that the end user would be
interested in any error messages regarding these issues.

The version number of currently installed app is read from "Assembly
informational version", and can be set by creating git tags. See README
for more information.