CLI changes + expansion, add better project path handling
Implemented the `List` subcommand to allow for users to query TCLI data in a (somewhat) simple way.
- `tcli list platforms` lists TCLI supported platforms for the specified OS.
- `tcli list registered-games` lists games registered for use within TCLI.
- `tcli list supported-games` lists games that TCLI supports.

Renamed the `Schema` command to `UpdateSchema`, expanded functionality to update the local schema upon invocation.

Implemented `ProjectPath`; a struct which *should* make dealing with project paths a bit easier. It simplifies the given path down to the root-most directory of the project, does some extra validation, and can even create directories if needed. Fun!

Added the `JsonParserError` so we stop unwrapping serde_json calls. Also tweaked the spelling of a couple comments.

Other stuff that I've probably forgotten about.