improve package installer and project create functions
The package installer setup code now uses the newly created package::error type, of which allows error states to be more expressive and explanatory.

The changes to the Project struct are a part of a continued effort to gate off raw project access. This ensures that we don't need to use a janky ProjectPath type to ensure that we know what goes where, etc.

This commit also includes a code formatting pass, however the changes are relatively minor.